Custom casing of table/column/function/view names
This adds support for defining the casing to be used in completions (for unquoted names). The casing is determined by the casing file, which consists of a \n-separated list of names using the preferred casing. This file is initially populated based on usage in SQL and PLPGSQL database functions but can then be modified by the user.