For 'JOIN <cursor>', suggest 'foo on foo.fooid = bar.fooid'
This is based on my previous work on suggesting join conditions, but here instead we suggest the whole join. What we do is simply check all the tables in the statement for FK relationships and then suggest joins based on those.

I think this will not only save key presses but also be rather useful when exploring an unknown (part of a) database.