Add landscape.io config file
I disabled the rule enforcing line breaks after binary operator, as it's not in accordance with the current version of PEP8 (which suggests the opposite: line breaks before binary operators).
I also disabled the too-many-arguments and too-many-locals rules, as they weren't producing anything useful.
And I disabled some pyflakes rules that are duplicates of pylint rules. The benefit of keeping the pylint rule rather than the pyflakes rule is that in the case of false positives, a single pylint rule can be disabled for a single line of code.